requirements.txt
# Orchestration & Logic
dspy
langchain 
langchain-text-splitters
pydantic-settings

# Document Parsing
docling

# Vector Store & Embeddings
faiss-cpu
sentence-transformers
rank_bm25

# Utilities
tqdm
requests
torch
bitsandbytes
